# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/translations/jobs": {
      "post": {
        "description": "Creates the PostgreSQL translation job and immediately enqueues it for background execution through BullMQ.",
        "operationId": "TranslationJobsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTranslationJobDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Translation job created and queued successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationJobDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create and enqueue a translation job",
        "tags": [
          "Translation Jobs"
        ]
      },
      "get": {
        "operationId": "TranslationJobsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TranslationJobDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List translation jobs",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/{id}/queue": {
      "get": {
        "operationId": "TranslationJobsController_queueJobStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get BullMQ status for a translation job owned by the authenticated shop",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/usage/summary": {
      "get": {
        "description": "Returns raw provider usage, estimated configured cost, provider/stage breakdowns and daily/monthly aggregates for the authenticated shop.",
        "operationId": "TranslationJobsController_usageSummary",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get translation usage and cost summary",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/usage/events": {
      "get": {
        "description": "Returns the most recent translation and AI-review provider attempts for the authenticated shop.",
        "operationId": "TranslationJobsController_usageEvents",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List recent translation usage events",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/{id}": {
      "get": {
        "operationId": "TranslationJobsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationJobDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get translation job details",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/{id}/execute": {
      "post": {
        "description": "Development/manual execution endpoint. Production jobs are normally executed by the BullMQ background worker. Duplicate execution is protected by the atomic database claim.",
        "operationId": "TranslationJobsController_execute",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute a queued translation job immediately",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/{id}/cancel": {
      "post": {
        "description": "Cancels the database lifecycle and removes a waiting BullMQ job when possible. Active work is cancelled cooperatively by the translation executor.",
        "operationId": "TranslationJobsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a queued or running translation job",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/{id}/retry-failed": {
      "post": {
        "description": "Requeues only FAILED items in PostgreSQL and automatically schedules the job again in BullMQ.",
        "operationId": "TranslationJobsController_retryFailed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retry failed translation items",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/{id}/resume": {
      "post": {
        "description": "Requeues interrupted GENERATING items in PostgreSQL and schedules the resumed job in BullMQ. RUNNING jobs must be stale for at least five minutes.",
        "operationId": "TranslationJobsController_resume",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resume a cancelled or stale running translation job",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/{id}/review": {
      "post": {
        "description": "Runs the snapshotted AI reviewer without overwriting the generated translation.",
        "operationId": "TranslationJobsController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Run AI review for a translation job",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/{jobId}/items/{itemId}/approve": {
      "post": {
        "description": "Supports approve-as-is and edit-and-approve. The final value is deterministically validated.",
        "operationId": "TranslationJobsController_approveItem",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveTranslationItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve a translation item",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/{jobId}/items/{itemId}/reject": {
      "post": {
        "operationId": "TranslationJobsController_rejectItem",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectTranslationItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject a translation item",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/{jobId}/items/{itemId}/regenerate": {
      "post": {
        "description": "Creates a new one-item translation job while preserving the original translation item and its audit history.",
        "operationId": "TranslationJobsController_regenerateItem",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Regenerate a translation item",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/{jobId}/items/{itemId}/publication/prepare": {
      "post": {
        "description": "Validates publication eligibility and returns the Shopify resource, key, locale, approved value and source digest required by the React Router Shopify gateway.",
        "operationId": "TranslationJobsController_preparePublication",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Translation publication prepared successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Prepare an approved translation for Shopify publication",
        "tags": [
          "Translation Jobs"
        ]
      }
    },
    "/api/v1/translations/jobs/{jobId}/items/{itemId}/publication/confirm": {
      "post": {
        "description": "Marks an approved translation as published only after Shopify translationsRegister succeeds and updates the local TranslationState projection.",
        "operationId": "TranslationJobsController_confirmPublication",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Translation publication confirmed successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm successful Shopify publication",
        "tags": [
          "Translation Jobs"
        ]
      }
    }
  },
  "info": {
    "title": "ACA Locale API",
    "description": "Technical API documentation for ACA Locale.\n\nACA Locale is an AI-powered localization management platform.\n\nThis specification is generated automatically from the NestJS backend.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTranslationJobDto": {
        "type": "object",
        "properties": {
          "targetLocale": {
            "type": "string",
            "description": "Target Shopify locale.",
            "example": "it"
          },
          "resourceTypes": {
            "description": "Shopify resource types to include in the translation job.",
            "example": [
              "PRODUCT",
              "COLLECTION",
              "PAGE"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxItems": {
            "type": "number",
            "description": "Maximum number of translation fields to include in this job.",
            "example": 100,
            "default": 100,
            "minimum": 1,
            "maximum": 500
          }
        },
        "required": [
          "targetLocale",
          "resourceTypes"
        ]
      },
      "TranslationJobResourceDto": {
        "type": "object",
        "properties": {
          "resourceType": {
            "type": "string",
            "example": "PRODUCT"
          },
          "shopifyResourceId": {
            "type": "string",
            "example": "gid://shopify/Product/123456789"
          }
        },
        "required": [
          "resourceType",
          "shopifyResourceId"
        ]
      },
      "TranslationJobFieldDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "title"
          },
          "type": {
            "type": "string",
            "example": "STRING"
          },
          "sourceLocale": {
            "type": "string",
            "example": "en"
          },
          "resource": {
            "$ref": "#/components/schemas/TranslationJobResourceDto"
          }
        },
        "required": [
          "key",
          "type",
          "sourceLocale",
          "resource"
        ]
      },
      "TranslationJobItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fieldId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "GENERATING",
              "GENERATED",
              "VALIDATED",
              "NEEDS_REVIEW",
              "APPROVED",
              "PUBLISHED",
              "FAILED"
            ],
            "example": "PENDING"
          },
          "sourceDigest": {
            "type": "string"
          },
          "sourceValue": {
            "type": "string",
            "example": "Example product title"
          },
          "translatedValue": {
            "type": "object",
            "nullable": true
          },
          "provider": {
            "type": "object",
            "nullable": true
          },
          "model": {
            "type": "object",
            "nullable": true
          },
          "generatedAt": {
            "type": "object",
            "nullable": true
          },
          "approvedAt": {
            "type": "object",
            "nullable": true
          },
          "publishedAt": {
            "type": "object",
            "nullable": true
          },
          "errorMessage": {
            "type": "object",
            "nullable": true
          },
          "field": {
            "$ref": "#/components/schemas/TranslationJobFieldDto"
          }
        },
        "required": [
          "id",
          "fieldId",
          "status",
          "sourceDigest",
          "sourceValue",
          "field"
        ]
      },
      "TranslationJobDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmv1234567890"
          },
          "sourceLocale": {
            "type": "string",
            "example": "en"
          },
          "targetLocale": {
            "type": "string",
            "example": "it"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "COMPLETED",
              "PARTIAL",
              "FAILED",
              "CANCELLED"
            ],
            "example": "QUEUED"
          },
          "provider": {
            "type": "string",
            "enum": [
              "OPENAI",
              "ANTHROPIC",
              "GOOGLE"
            ],
            "example": "OPENAI",
            "nullable": true
          },
          "model": {
            "type": "object",
            "example": "gpt-5.6-luna",
            "nullable": true
          },
          "totalItems": {
            "type": "number",
            "example": 100
          },
          "completedItems": {
            "type": "number",
            "example": 0
          },
          "failedItems": {
            "type": "number",
            "example": 0
          },
          "startedAt": {
            "type": "object",
            "nullable": true
          },
          "completedAt": {
            "type": "object",
            "nullable": true
          },
          "errorMessage": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationJobItemDto"
            }
          }
        },
        "required": [
          "id",
          "sourceLocale",
          "targetLocale",
          "status",
          "totalItems",
          "completedItems",
          "failedItems",
          "createdAt",
          "updatedAt",
          "items"
        ]
      },
      "TranslationJobDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmv1234567890"
          },
          "sourceLocale": {
            "type": "string",
            "example": "en"
          },
          "targetLocale": {
            "type": "string",
            "example": "it"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "COMPLETED",
              "PARTIAL",
              "FAILED",
              "CANCELLED"
            ],
            "example": "QUEUED"
          },
          "provider": {
            "type": "string",
            "enum": [
              "OPENAI",
              "ANTHROPIC",
              "GOOGLE"
            ],
            "example": "OPENAI",
            "nullable": true
          },
          "model": {
            "type": "object",
            "example": "gpt-5.6-luna",
            "nullable": true
          },
          "totalItems": {
            "type": "number",
            "example": 100
          },
          "completedItems": {
            "type": "number",
            "example": 0
          },
          "failedItems": {
            "type": "number",
            "example": 0
          },
          "startedAt": {
            "type": "object",
            "nullable": true
          },
          "completedAt": {
            "type": "object",
            "nullable": true
          },
          "errorMessage": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sourceLocale",
          "targetLocale",
          "status",
          "totalItems",
          "completedItems",
          "failedItems",
          "createdAt",
          "updatedAt"
        ]
      },
      "ApproveTranslationItemDto": {
        "type": "object",
        "properties": {
          "approvedValue": {
            "type": "string",
            "description": "Optional manually edited translation. If omitted, the generated translatedValue is approved as-is.",
            "example": "Snowboard con inventario non monitorato"
          },
          "note": {
            "type": "string",
            "description": "Optional human review note.",
            "example": "Terminology adjusted for the Italian storefront."
          }
        }
      },
      "RejectTranslationItemDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "description": "Reason why the generated translation was rejected.",
            "example": "Terminology is not suitable for the Italian market."
          }
        },
        "required": [
          "note"
        ]
      }
    }
  }
}